mergen v0.0.1 Mergen

The high-level API.

Functions

due_for_lesson?(item)

due_for_review?(item)

export_json(filepath \\ "")

export_json(String.t()) :: {:ok, integer(), String.t()} | {:error, String.t()}

Exports all items in the database to a JSON file.

If successful, returns the number of exported items and the export path. Otherwise, returns an error string.

import_json(filepath)

import_json(String.t()) :: {:ok, String.t()} | {:error, String.t()}

Imports a list of items from a JSON file into the database.

Currently does not check for duplicates, and last_reviewed values are ignored.

randomized_lesson_queue()

randomized_review_queue()

submit_answer(item_id, answer)

submit_answer(integer(), String.t()) :: Mergen.Items.Item.t()